proper merge
